+++ hitman_codename_47/gui.cpp
@@ -7,6 +7,7 @@
 #include "../hitman_common/gui.hpp"
 
 const std::vector<std::string> map_names = {
+    "",
     "#1 Training",
     "#2 Kowloon Triads in Gang War",
     "#3 Ambush at the Wang Fou Restaurant",
@@ -29,18 +30,18 @@ GameGui hitman_codename_47::gui(
         auto game_name = settings.show_game_version
                              ? std::format("{} [{}]", GAME_NAME, version)
                              : GAME_NAME;
-        hitman_common::gui(
+        hitman_common::gui_header(
             settings,
             fonts,
-            1.0f,
-            stats,
             game_name,
             stats.difficulty == 0   ? "Easy"
             : stats.difficulty == 1 ? "Normal"
             : stats.difficulty == 2 ? "Hard"
                                     : "",
-            map_names,
-            {}
+            map_names[stats.map],
+            stats.map,
+            stats.map_stage,
+            stats.time
         );
     };
 }