+++ hitman_contracts/gui.cpp
@@ -10,6 +10,7 @@
 #include "../imgui_utils.hpp"
 
 const std::vector<std::string> map_names = {
+    "",
     "#1 Asylum Aftermath",
     "#2 The Meat King's Party",
     "#3 The Bjarkhov Bomb",
@@ -31,30 +32,37 @@ GameGui hitman_contracts::gui(
         auto game_name = settings.show_game_version
                              ? std::format("{} [{}]", GAME_NAME, version)
                              : GAME_NAME;
-        hitman_common::gui(
+        hitman_common::gui_header(
             settings,
             fonts,
-            1.0f,
-            stats,
             game_name,
             stats.difficulty == 1   ? "Normal"
             : stats.difficulty == 2 ? "Expert"
             : stats.difficulty == 3 ? "Professional"
                                     : "",
-            map_names,
-            {
-                {"Close Encounters", stats.close_encounters},
-                {"Alerts", stats.alerts},
-                {"Shots Fired", stats.shots_fired},
-                {"Headshots", stats.headshots},
-                {"Enemies Killed", stats.enemies_killed},
-                {"Enemies Wounded", stats.enemies_wounded},
-                {"Innocents Killed", stats.innocents_killed},
-                {"Innocents Wounded", stats.innocents_wounded},
-            }
+            map_names[stats.map],
+            stats.map,
+            stats.map_stage,
+            stats.time
         );
         if (stats.map > 0) {
-            ImGui::Spacing();
+            hitman_common::gui_table(
+                settings,
+                fonts,
+                stats.rating,
+                stats.map,
+                stats.map_stage,
+                {
+                    {"Close Encounters", stats.close_encounters},
+                    {"Alerts", stats.alerts},
+                    {"Shots Fired", stats.shots_fired},
+                    {"Headshots", stats.headshots},
+                    {"Enemies Killed", stats.enemies_killed},
+                    {"Enemies Wounded", stats.enemies_wounded},
+                    {"Innocents Killed", stats.innocents_killed},
+                    {"Innocents Wounded", stats.innocents_wounded},
+                }
+            );
             ImGui::BeginTable(
                 "Statistics",
                 2,