+++ stats.hpp
@@ -14,8 +14,12 @@ struct StatsValue {
 
 StatsValue<int32_t> stats_value(int32_t value, bool required = true);
 
+// only for HMA
+enum class CheckpointType { UNRATED, NO_TARGETS, TARGETS };
+
 enum class MapStage { pre, main, post };
 
+// TODO split this up per game, use std::any for type erasure
 struct Stats {
     float time;
     int map;
@@ -45,6 +49,7 @@ struct Stats {
     StatsValue<int32_t> suit_left;          // bm
     int32_t shots_hit;                      // bm
     int32_t accident_kills;                 // bm
+    CheckpointType checkpoint_type;         // hma: checkpoint type needed for display
     int32_t score_for_max_rating;           // hma: score needed for max rating
     int32_t score_total;                    // hma: actual score
     std::string score_rating;               // hma: actual rating